Split scanner risk and blast-radius helpers
