+++ src/agent_bom/scanners/__init__.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 import asyncio
 import logging
-import math
 import threading
 import time
 from typing import Any, Optional
@@ -45,6 +44,8 @@ from agent_bom.owasp import tag_blast_radius
 from agent_bom.owasp_agentic import tag_blast_radius as tag_owasp_agentic
 from agent_bom.owasp_mcp import tag_blast_radius as tag_owasp_mcp
 from agent_bom.package_utils import normalize_package_name
+from agent_bom.scanners.blast_radius import _HOP_RISK_FACTORS, expand_blast_radius_hops
+from agent_bom.scanners.risk import _parse_cvss4_vector, cvss_to_severity, parse_cvss_vector, parse_osv_severity
 from agent_bom.soc2 import tag_blast_radius as tag_soc2
 from agent_bom.vuln_compliance import tag_vulnerability as _tag_vuln
 
@@ -293,227 +294,6 @@ def _get_scan_cache():  # noqa: ANN202
                     record_scan_warning("scan cache unavailable")
                     _scan_cache_instance = False  # mark as attempted, don't retry
     return _scan_cache_instance if _scan_cache_instance is not False else None
-
-
-# Map CVSS scores to severity
-def cvss_to_severity(score: Optional[float]) -> Severity:
-    if score is None:
-        return Severity.UNKNOWN
-    if score >= 9.0:
-        return Severity.CRITICAL
-    if score >= 7.0:
-        return Severity.HIGH
-    if score >= 4.0:
-        return Severity.MEDIUM
-    if score > 0:
-        return Severity.LOW
-    return Severity.NONE
-
-
-# CVSS 3.1 Base Score metric weights.
-# Reference: FIRST CVSS v3.1 Specification, Section 7.4 — Metric Values
-# https://www.first.org/cvss/v3.1/specification-document#7-4-Metric-Values
-_CVSS3_AV = {"N": 0.85, "A": 0.62, "L": 0.55, "P": 0.20}  # Attack Vector
-_CVSS3_AC = {"L": 0.77, "H": 0.44}  # Attack Complexity
-_CVSS3_PR_U = {"N": 0.85, "L": 0.62, "H": 0.27}  # Privileges Required (Scope Unchanged)
-_CVSS3_PR_C = {"N": 0.85, "L": 0.68, "H": 0.50}  # Privileges Required (Scope Changed)
-_CVSS3_UI = {"N": 0.85, "R": 0.62}  # User Interaction
-_CVSS3_CIA = {"N": 0.00, "L": 0.22, "H": 0.56}  # Confidentiality / Integrity / Availability
-
-
-def _parse_cvss4_vector(vector: str) -> Optional[float]:
-    """Extract an approximate base score from a CVSS 4.0 vector string.
-
-    CVSS v4.0 scoring requires a complex lookup-table algorithm that isn't
-    practical to reimplement inline (700+ macro-vector combinations).  Instead
-    we estimate a base score from the *impact* and *exploitability* metric
-    values using a simplified weighted model that tracks the official
-    calculator within ±0.5 for typical vectors.
-
-    Returns ``None`` if the vector cannot be parsed.
-    """
-    try:
-        parts = vector.split("/")[1:]
-        m = dict(p.split(":") for p in parts)
-
-        # Attack Vector / Complexity / Privileges / User Interaction
-        av = {"N": 1.0, "A": 0.75, "L": 0.55, "P": 0.20}.get(m.get("AV", ""), None)
-        ac = {"L": 1.0, "H": 0.55}.get(m.get("AC", ""), None)
-        at = {"N": 1.0, "P": 0.60}.get(m.get("AT", ""), None)  # Attack Requirements
-        pr = {"N": 1.0, "L": 0.65, "H": 0.30}.get(m.get("PR", ""), None)
-        ui = {"N": 1.0, "P": 0.70, "A": 0.55}.get(m.get("UI", ""), None)
-
-        # Vulnerable-system impact
-        vc = {"H": 0.56, "L": 0.22, "N": 0.0}.get(m.get("VC", ""), None)
-        vi = {"H": 0.56, "L": 0.22, "N": 0.0}.get(m.get("VI", ""), None)
-        va = {"H": 0.56, "L": 0.22, "N": 0.0}.get(m.get("VA", ""), None)
-
-        required = (av, ac, at, pr, ui, vc, vi, va)
-        if any(v is None for v in required):
-            return None
-
-        # Type narrowing for mypy — the None case is already handled by the
-        # ``any(v is None ...)`` guard above; these are runtime-safe casts.
-        av, ac, at, pr, ui = float(av), float(ac), float(at), float(pr), float(ui)  # type: ignore[arg-type]
-        vc, vi, va = float(vc), float(vi), float(va)  # type: ignore[arg-type]
-
-        # Subsequent-system impact (optional — defaults to None=0)
-        sc = {"H": 0.56, "L": 0.22, "N": 0.0}.get(m.get("SC", "N"), 0.0)
-        si = {"H": 0.56, "L": 0.22, "N": 0.0}.get(m.get("SI", "N"), 0.0)
-        sa = {"H": 0.56, "L": 0.22, "N": 0.0}.get(m.get("SA", "N"), 0.0)
-
-        isc = 1.0 - (1.0 - vc) * (1.0 - vi) * (1.0 - va)
-        isc_sub = 1.0 - (1.0 - sc) * (1.0 - si) * (1.0 - sa)
-        impact = max(isc, isc + 0.25 * isc_sub)  # Subsequent amplifies
-
-        if impact <= 0:
-            return 0.0
-
-        exploit = av * ac * at * pr * ui
-        raw = min(10.0, 1.1 * (6.42 * impact + 8.22 * exploit * 0.6))
-
-        return math.ceil(raw * 10) / 10.0
-    except Exception as exc:
-        _logger.debug("CVSS 4.0 vector parse failed for %r: %s", vector, exc)
-        return None
-
-
-def parse_cvss_vector(vector: str) -> Optional[float]:
-    """Compute CVSS base score from a vector string (v3.x and v4.0).
-
-    Examples:
-        'CVSS:3.1/AV:N/AC:L/PR:N/UI:N/S:U/C:H/I:H/A:H' → 9.8
-        'CVSS:4.0/AV:N/AC:L/AT:N/PR:N/UI:N/VC:H/VI:H/VA:H/SC:N/SI:N/SA:N' → ~9.3
-    """
-    try:
-        if vector.startswith("CVSS:4"):
-            return _parse_cvss4_vector(vector)
-        if not vector.startswith("CVSS:3"):
-            return None
-        # Strip prefix
-        parts = vector.split("/")[1:]
-        metrics = dict(p.split(":") for p in parts)
-
-        av = _CVSS3_AV.get(metrics.get("AV", ""), None)
-        ac = _CVSS3_AC.get(metrics.get("AC", ""), None)
-        scope = metrics.get("S", "U")
-        pr_map = _CVSS3_PR_C if scope == "C" else _CVSS3_PR_U
-        pr = pr_map.get(metrics.get("PR", ""), None)
-        ui = _CVSS3_UI.get(metrics.get("UI", ""), None)
-        c = _CVSS3_CIA.get(metrics.get("C", ""), None)
-        i = _CVSS3_CIA.get(metrics.get("I", ""), None)
-        a = _CVSS3_CIA.get(metrics.get("A", ""), None)
-
-        if any(v is None for v in (av, ac, pr, ui, c, i, a)):
-            return None
-
-        # Type narrowing for mypy — the None case is already handled by the
-        # ``any(v is None ...)`` guard above; these are runtime-safe casts.
-        av, ac, pr, ui = float(av), float(ac), float(pr), float(ui)  # type: ignore[arg-type]
-        c, i, a = float(c), float(i), float(a)  # type: ignore[arg-type]
-
-        isc_base = 1.0 - (1.0 - c) * (1.0 - i) * (1.0 - a)
-        if scope == "C":
-            isc = 7.52 * (isc_base - 0.029) - 3.25 * ((isc_base - 0.02) ** 15)
-        else:
-            isc = 6.42 * isc_base
-
-        if isc <= 0:
-            return 0.0
-
-        exploitability = 8.22 * av * ac * pr * ui
-
-        if scope == "C":
-            raw = min(1.08 * (isc + exploitability), 10.0)
-        else:
-            raw = min(isc + exploitability, 10.0)
-
-        # Roundup to one decimal (CVSS spec: ceiling to 1 decimal)
-        return math.ceil(raw * 10) / 10.0
-
-    except Exception as exc:
-        _logger.debug("CVSS vector parse failed for %r: %s", vector, exc)
-        return None
-
-
-def parse_osv_severity(vuln_data: dict) -> tuple[Severity, Optional[float], Optional[str]]:
-    """Extract severity, CVSS score, and severity source from OSV vulnerability data.
-
-    Returns ``(severity, cvss_score, severity_source)`` where *severity_source*
-    indicates where the severity was derived from:
-
-    - ``"cvss"`` — parsed from a CVSS v3/v4 vector or numeric score
-    - ``"osv_database"`` — from ``database_specific.severity``
-    - ``"osv_ecosystem"`` — from ``ecosystem_specific.severity``
-    - ``"ghsa_heuristic"`` — inferred MEDIUM for reviewed GHSA advisories
-    - ``None`` — severity is UNKNOWN, no source available
-    """
-    cvss_score = None
-    severity = Severity.UNKNOWN  # Default — never silently inflate to MEDIUM
-    severity_source: Optional[str] = None
-
-    # Check severity array — may be numeric score or CVSS vector string
-    for sev in vuln_data.get("severity", []):
-        if sev.get("type") in ("CVSS_V3", "CVSS_V3_1", "CVSS_V4"):
-            score_str = sev.get("score", "")
-            try:
-                parsed = float(score_str)
-                # CVSS scores must be 0.0–10.0
-                if 0.0 <= parsed <= 10.0:
-                    cvss_score = parsed
-            except ValueError:
-                # It's a CVSS vector string — compute the base score
-                computed = parse_cvss_vector(score_str)
-                if computed is not None and 0.0 <= computed <= 10.0:
-                    cvss_score = computed
-
-    # Check database_specific for severity label (reliable fallback)
-    db_specific = vuln_data.get("database_specific", {})
-    if "severity" in db_specific:
-        sev_str = db_specific["severity"].upper()
-        severity_map = {
-            "CRITICAL": Severity.CRITICAL,
-            "HIGH": Severity.HIGH,
-            "MODERATE": Severity.MEDIUM,
-            "MEDIUM": Severity.MEDIUM,
-            "LOW": Severity.LOW,
-        }
-        resolved = severity_map.get(sev_str, Severity.UNKNOWN)
-        if resolved != Severity.UNKNOWN:
-            severity = resolved
-            severity_source = "osv_database"
-
-    # CVSS score overrides label-based severity
-    if cvss_score is not None:
-        severity = cvss_to_severity(cvss_score)
-        severity_source = "cvss"
-
-    # If still UNKNOWN, try to infer from ecosystem_specific or affected data
-    if severity == Severity.UNKNOWN:
-        eco_specific = vuln_data.get("ecosystem_specific", {})
-        if isinstance(eco_specific, dict) and "severity" in eco_specific:
-            sev_str = str(eco_specific["severity"]).upper()
-            severity_map = {
-                "CRITICAL": Severity.CRITICAL,
-                "HIGH": Severity.HIGH,
-                "MODERATE": Severity.MEDIUM,
-                "MEDIUM": Severity.MEDIUM,
-                "LOW": Severity.LOW,
-            }
-            resolved = severity_map.get(sev_str, severity)
-            if resolved != Severity.UNKNOWN:
-                severity = resolved
-                severity_source = "osv_ecosystem"
-
-    # Last resort: if GHSA advisory, treat as at least MEDIUM
-    # (GHSA advisories are reviewed and vetted — they wouldn't exist without merit)
-    if severity == Severity.UNKNOWN:
-        vuln_id = vuln_data.get("id", "")
-        if vuln_id.startswith("GHSA-"):
-            severity = Severity.MEDIUM
-            severity_source = "ghsa_heuristic"
-
-    return severity, cvss_score, severity_source
 
 
 def _candidate_package_names(package_name: str, ecosystem: str = "", source_package: str | None = None) -> set[str]:
@@ -1942,120 +1722,6 @@ async def scan_agents_with_enrichment(
         blast_radii.sort(key=lambda br: br.risk_score, reverse=True)
 
     return blast_radii
-
-
-# ── Multi-hop risk amplification factors per hop distance ─────────────────
-_HOP_RISK_FACTORS: dict[int, float] = {
-    1: 1.0,
-    2: 0.7,
-    3: 0.5,
-    4: 0.35,
-    5: 0.25,
-}
-
-
-def expand_blast_radius_hops(
-    blast_radii: list[BlastRadius],
-    agents: list[Agent],
-    max_depth: int = 1,
-) -> None:
-    """Expand blast radii with multi-hop delegation chain analysis.
-
-    For each blast radius, traces agent→server→agent delegation chains
-    beyond the initial (1-hop) affected agents. Uses BFS with cycle
-    detection to avoid infinite loops.
-
-    When ``max_depth`` is 1 (default), this is a no-op — zero overhead.
-
-    Args:
-        blast_radii: Existing 1-hop blast radius findings (mutated in place).
-        agents: All discovered agents (for cross-referencing servers).
-        max_depth: Maximum hop depth (1-5). Clamped to [1, 5].
-    """
-    max_depth = max(1, min(max_depth, 5))
-    if max_depth <= 1:
-        return
-
-    # Build lookup: server_name → list of agents that use that server
-    server_to_agents: dict[str, list[Agent]] = {}
-    for agent in agents:
-        for server in agent.mcp_servers:
-            server_to_agents.setdefault(server.name, []).append(agent)
-
-    # Build lookup: agent_name → list of server names it uses
-    agent_to_servers: dict[str, list[str]] = {}
-    for agent in agents:
-        agent_to_servers[agent.name] = [s.name for s in agent.mcp_servers]
-
-    for br in blast_radii:
-        direct_agent_names = {a.name for a in br.affected_agents}
-        direct_server_names = {s.name for s in br.affected_servers}
-
-        # BFS: queue items are (agent_name, current_hop, chain_so_far)
-        visited_agents: set[str] = set(direct_agent_names)
-        visited_servers: set[str] = set(direct_server_names)
-        transitive_agents: list[dict] = []
-        transitive_creds: list[str] = []
-        chains: list[str] = []
-
-        # Seed BFS from direct agents
-        queue: list[tuple[str, int, list[str]]] = []
-        for agent in br.affected_agents:
-            for srv_name in agent_to_servers.get(agent.name, []):
-                if srv_name not in direct_server_names:
-                    queue.append((agent.name, 1, [agent.name, srv_name]))
-                    visited_servers.add(srv_name)
-
-        max_hop_reached = 1
-        while queue:
-            agent_name, hop, chain = queue.pop(0)
-            if hop >= max_depth:
-                continue
-
-            # The last element in chain is a server name — find agents on it
-            current_server = chain[-1]
-            for next_agent in server_to_agents.get(current_server, []):
-                if next_agent.name in visited_agents:
-                    continue
-                visited_agents.add(next_agent.name)
-                next_hop = hop + 1
-                max_hop_reached = max(max_hop_reached, next_hop)
-
-                new_chain = chain + [next_agent.name]
-                chain_str = "\u2192".join(new_chain)
-                chains.append(chain_str)
-
-                # Collect transitive agent info
-                agent_creds = []
-                for srv in next_agent.mcp_servers:
-                    agent_creds.extend(srv.credential_names)
-                agent_creds = list(set(agent_creds))
-
-                transitive_agents.append(
-                    {
-                        "name": next_agent.name,
-                        "type": next_agent.agent_type.value,
-                        "hop": next_hop,
-                        "chain": chain_str,
-                    }
-                )
-                transitive_creds.extend(agent_creds)
-
-                # Continue BFS: look at servers this agent connects to
-                if next_hop < max_depth:
-                    for srv_name in agent_to_servers.get(next_agent.name, []):
-                        if srv_name not in visited_servers:
-                            visited_servers.add(srv_name)
-                            queue.append((next_agent.name, next_hop, new_chain + [srv_name]))
-
-        if transitive_agents:
-            br.hop_depth = max_hop_reached
-            br.delegation_chain = chains
-            br.transitive_agents = transitive_agents
-            br.transitive_credentials = list(set(transitive_creds))
-            # Transitive risk = base risk * hop factor
-            factor = _HOP_RISK_FACTORS.get(max_hop_reached, 0.25)
-            br.transitive_risk_score = round(br.risk_score * factor, 2)
 
 
 def scan_agents_sync(
